# openapi.yaml
openapi: 3.1.0
info:
  title: Powerlifting App API
  version: 1.0.0
  description: |
    REST API for the Powerlifting App — a multi-user workout tracking platform
    for athletes and coaches.

    ## Auth
    All endpoints except `/auth/register`, `/auth/login`, and `/auth/token`
    require `Authorization: Bearer <access_token>`.

    ## Errors
    All error responses use **RFC 7807 Problem Details** with
    `Content-Type: application/problem+json`.

    ## REQ mapping
    Each operation carries an `x-req` extension listing the requirement IDs it
    fulfils (e.g. `x-req: [REQ-001]`).
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# =============================================================================
# COMPONENTS
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    ISODateTime:
      type: string
      format: date-time
      example: "2026-03-11T14:00:00Z"
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: "https://api.example.com/problems/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "email must be a valid email address"
        instance:
          type: string
          format: uri
          example: "/api/v1/auth/register"

    # --- Auth (REQ-001, REQ-002) ----------------------------------------------
...
    UserSummary:
      type: object
      required: [id, email, role, profile]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        role:
          type: string
          enum: [athlete, coach]
        profile:
          oneOf:
            - $ref: '#/components/schemas/AthleteProfile'
            - $ref: '#/components/schemas/CoachProfile'
          discriminator:
            propertyName: role
    AthleteProfile:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          example: "Jane Lifter"
        weightClass:
          type: string
          nullable: true
          example: "72kg"
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: "1995-06-15"
    CoachProfile:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          example: "Coach Mike"
        bio:
          type: string
          nullable: true
          example: "USAPL certified coach, 10 years experience."
    UpdateAthleteProfileRequest:
      type: object
      minProperties: 1
      properties:
        displayName:
          type: string
        weightClass:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
          nullable: true
    UpdateCoachProfileRequest:
      type: object
      minProperties: 1
      properties:
        displayName:
          type: string
        bio:
          type: string
          nullable: true

    # --- Relationships (REQ-005) ----------------------------------------------
...
    Set:
      type: object
      required: [id, sessionId, exerciseId, weightLbs, reps, order, isPR, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        sessionId:
          $ref: '#/components/schemas/UUID'
        exerciseId:
          $ref: '#/components/schemas/UUID'
        weightLbs:
          type: number
          format: float
          minimum: 0
          example: 225.0
        reps:
          type: integer
          minimum: 1
          example: 5
        rpe:
          type: number
          format: float
          minimum: 1
          maximum: 10
          nullable: true
          example: 8.5
        rir:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
          example: 2
        videoUrl:
          type: string
          format: uri
          nullable: true
        notes:
          type: string
          nullable: true
        order:
          type: integer
          minimum: 0
        isPR:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
        updatedAt:
          $ref: '#/components/schemas/ISODateTime'
    CreateSetRequest:
      type: object
      required: [exerciseId, weightLbs, reps]
      properties:
        exerciseId:
          $ref: '#/components/schemas/UUID'
        weightLbs:
          type: number
          format: float
          minimum: 0
        reps:
          type: integer
          minimum: 1
        rpe:
          type: number
          format: float
          minimum: 1
          maximum: 10
          nullable: true
        rir:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        videoUrl:
          type: string
          format: uri
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateSetRequest:
      type: object
      minProperties: 1
      properties:
        weightLbs:
          type: number
          format: float
          minimum: 0
        reps:
          type: integer
          minimum: 1
        rpe:
          type: number
          format: float
          minimum: 1
          maximum: 10
          nullable: true
        rir:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        videoUrl:
          type: string
          format: uri
          nullable: true
        notes:
          type: string
          nullable: true

    # --- Personal Records (REQ-018, REQ-019, REQ-020) ------------------------
...
  responses:
    400BadRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/validation-error"
            title: "Validation Error"
            status: 400
            detail: "password must be at least 8 characters"
    401Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Bearer token is missing or expired"
    403Forbidden:
      description: Authenticated but insufficient role or resource ownership.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/forbidden"
            title: "Forbidden"
            status: 403
            detail: "You do not have access to this resource"
    404NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/not-found"
            title: "Not Found"
            status: 404
...
    422UnprocessableEntity:
      description: Business rule violation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/unprocessable"
            title: "Unprocessable Entity"
            status: 422
            detail: "Sets cannot be edited after the session is closed"

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /users/{userId}:
    get:
      tags: [Users]
      operationId: getUserProfile
      summary: Get a user's profile
      x-req: [REQ-003, REQ-004]
      description: |
        **REQ-003, REQ-004** — Retrieves an athlete or coach profile. Athletes
        can read their own profile and profiles of their linked coaches. Coaches
        can read their own profile and profiles of their linked athletes.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: User profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'

    patch:
      tags: [Users]
      operationId: updateUserProfile
      summary: Update a user's profile
      x-req: [REQ-003, REQ-004]
      description: |
        **REQ-003, REQ-004** — Partial update of profile fields. Users may only
        update their own profile. Shape of accepted body depends on caller's
        role.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/UpdateAthleteProfileRequest'
                - $ref: '#/components/schemas/UpdateCoachProfileRequest'
      responses:
        "200":
          description: Profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'

  # ---------------------------------------------------------------------------
  # RELATIONSHIPS  ·  REQ-005
  # ---------------------------------------------------------------------------
...
  /sessions/{sessionId}/sets:
    post:
      tags: [Sets]
      operationId: createSet
      summary: Log a set in an open session
      x-req: [REQ-014]
      description: |
        **REQ-014** — Appends a set to the session (assigned `order = max + 1`).
        Required fields: `exerciseId`, `weightLbs`, `reps`. Returns 422 if the
        session is closed.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSetRequest'
      responses:
        "201":
          description: Set logged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Set'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "422":
          $ref: '#/components/responses/422UnprocessableEntity'
  /sessions/{sessionId}/sets/{setId}:
    patch:
      tags: [Sets]
      operationId: updateSet
      summary: Edit a set in an open session
      x-req: [REQ-015]
      description: |
        **REQ-015** — Partial update of any set field. Returns 422 if the
        parent session is `closed`.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: setId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSetRequest'
      responses:
        "200":
          description: Set updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Set'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "422":
          $ref: '#/components/responses/422UnprocessableEntity'

    delete:
      tags: [Sets]
      operationId: deleteSet
      summary: Delete a set from an open session
      x-req: [REQ-015]
      description: |
        **REQ-015** — Removes the set. Returns 422 if the parent session is
        `closed`.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: setId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Set deleted.
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "422":
          $ref: '#/components/responses/422UnprocessableEntity'

  # ---------------------------------------------------------------------------
  # PERSONAL RECORDS  ·  REQ-018, REQ-019, REQ-020
  # ---------------------------------------------------------------------------
...